Resolver primeiros Problemas: PerfilServlet (Passar user, de tabela foi para outras) MyChallsServlet (Passar Foto) IntialPageServlet (Adicionar Página Desafio e Servlet também)
